// src/types.ts
export type ConvertlyInput = Blob | ArrayBuffer | Uint8Array | Buffer | string;

export type ConvertlyFileInput = {
  file?: ConvertlyInput;
  sourceUrl?: string;
  filename?: string;
  contentType?: string;
};